Alteração de tipo, model shipment

+++ db/migrate/20171002170216_create_shipments.rb
@@ -1,7 +1,7 @@
 class CreateShipments < ActiveRecord::Migration[5.1]
   def change
     create_table :shipments do |t|
-      t.double :value
+      t.float :value
       t.integer :weight_kg
       t.references :contact, foreign_key: true
       t.references :origin, foreign_key: true
